Reduce to one single openhab setup function

+++ functions/openhab.sh
@@ -1,16 +1,45 @@
 #!/usr/bin/env bash
 
-openhab2_stable_setup() {
-  echo -n "$(timestamp) [openHABian] Installing openHAB 2.1 (stable)... "
-  echo "deb https://dl.bintray.com/openhab/apt-repo2 stable main" > /etc/apt/sources.list.d/openhab2.list
-  #echo "deb https://dl.bintray.com/openhab/apt-repo2 testing main" > /etc/apt/sources.list.d/openhab2.list
-  #echo "deb http://openhab.jfrog.io/openhab/openhab-linuxpkg unstable main" > /etc/apt/sources.list.d/openhab2.list
-  cond_redirect wget -O openhab-key.asc 'https://bintray.com/user/downloadSubjectPublicKey?username=openhab'
-  cond_redirect apt-key add openhab-key.asc
-  if [ $? -ne 0 ]; then echo "FAILED (key)"; exit 1; fi
-  rm -f openhab-key.asc
+openhab2_setup() {
+  introtext_stable="You are about to install or upgrade to the latest stable openHAB release.\\n
+Please be aware, that downgrading from a newer unstable snapshot build is not officially supported. Please consult with the documentation or community forum and be sure to take a full openHAB configuration backup first!"
+  successtext_stable="The stable release of openHAB is now installed on your system. Please test the correct behavior of your setup. You might need to adapt your configuration, if available. Check the \"openHAB Release Notes\" and the official announcements to learn about additons, fixes and changes."
+  # TODO: Add back in when links are known
+  # ➡ http://www.kaikreuzer.de/2017/06/28/openhab21
+  # ➡ https://github.com/openhab/openhab-distro/releases/tag/2.1.0"
+
+  introtext_unstable="You are about to switch over to the latest openHAB 2 unstable snapshot build. The daily snapshot builds contain the latest features and improvements but might also suffer from bugs or incompatibilities. Please be sure to take a full openHAB configuration backup first!"
+  successtext_unstable="The latest unstable snapshot build of openHAB 2 is now running on your system. Please test the correct behavior of your setup. You might need to adapt your configuration, if available.\\nIf you find any problem or bug, please report it and state the snapshot version you are on. To stay up-to-date with improvements and bug fixes you should upgrade your packages (the openhab2 package) regularly."
+
+  if [ "$1" == "unstable" ]; then
+    UNSTABLE=1
+  fi
+
+  if [ ! -n "$UNSTABLE" ]; then
+    echo -n "$(timestamp) [openHABian] Installing or upgrading to latest openHAB release (stable)... "
+    introtext=$introtext_stable
+    successtext=$successtext_stable
+  else
+    echo -n "$(timestamp) [openHABian] Installing or switching to latest openHAB snapshot (unstable)... "
+    introtext=$introtext_unstable
+    successtext=$successtext_unstable
+  fi
+
+  if [ -n "$INTERACTIVE" ]; then
+    if ! (whiptail --title "Description, Continue?" --yes-button "Continue" --no-button "Back" --yesno "$introtext" 15 80) then echo "CANCELED"; return 0; fi
+  fi
+
+  if [ ! -n "$UNSTABLE" ]; then
+    echo "deb https://dl.bintray.com/openhab/apt-repo2 stable main" > /etc/apt/sources.list.d/openhab2.list
+    cond_redirect wget -O openhab-key.asc 'https://bintray.com/user/downloadSubjectPublicKey?username=openhab'
+    cond_redirect apt-key add openhab-key.asc
+    if [ $? -ne 0 ]; then echo "FAILED (key)"; exit 1; fi
+    rm -f openhab-key.asc
+  else
+    echo "deb http://openhab.jfrog.io/openhab/openhab-linuxpkg unstable main" > /etc/apt/sources.list.d/openhab2.list
+  fi
   cond_redirect apt update
-  cond_redirect apt -y install openhab2
+  cond_redirect apt -y -o Dpkg::Options::="--force-confnew" install openhab2
   if [ $? -ne 0 ]; then echo "FAILED (apt)"; exit 1; fi
   cond_redirect adduser openhab dialout
   cond_redirect adduser openhab tty
@@ -24,82 +53,20 @@ openhab2_stable_setup() {
   if [ -n "$UNATTENDED" ]; then
     cond_redirect systemctl stop openhab2.service || true
   else
-    cond_redirect systemctl start openhab2.service || true
-  fi
-}
-
-openhab2_unstable_setup() {
-  introtext="You are about to switch over to the latest openHAB 2 unstable build. The daily snapshot builds contain the latest features and improvements but may also suffer from bugs or incompatibilities.
-If prompted if files should be replaced by newer ones, select Yes. Please be sure to take a full openHAB configuration backup first!"
-  successtext="The latest unstable/snapshot build of openHAB 2 is now running on your system. If already available, check the function of your configuration now. If you find any problem or bug, please report it and state the snapshot version you are on. To stay up-to-date with improvements and bug fixes you should upgrade your packages regularly."
-  echo -n "$(timestamp) [openHABian] Installing or switching to openHAB 2.2 SNAPSHOT (unstable)... "
-
-  if [ -n "$INTERACTIVE" ]; then
-    if ! (whiptail --title "Description, Continue?" --yes-button "Continue" --no-button "Back" --yesno "$introtext" 15 80) then echo "CANCELED"; return 0; fi
+    cond_redirect systemctl restart openhab2.service || true
   fi
 
-  echo "deb http://openhab.jfrog.io/openhab/openhab-linuxpkg unstable main" > /etc/apt/sources.list.d/openhab2.list
-  cond_redirect apt update
-  cond_redirect apt --yes --force-yes -o Dpkg::Options::="--force-confdef" -o Dpkg::Options::="--force-confnew" install openhab2
-  if [ $? -ne 0 ]; then echo "FAILED (apt)"; exit 1; fi
-  cond_redirect adduser openhab dialout
-  cond_redirect adduser openhab tty
-  cond_redirect adduser openhab gpio
-  cond_redirect adduser openhab audio
-  cond_redirect dashboard_add_tile openhabiandocs
-  if [ $? -ne 0 ]; then echo "FAILED (dashboard tile)"; fi
-  cond_redirect systemctl daemon-reload
-  cond_redirect systemctl enable openhab2.service
-  if [ $? -eq 0 ]; then echo "OK"; else echo "FAILED"; exit 1; fi
-  cond_redirect systemctl restart openhab2.service || true
-
   if [ -n "$INTERACTIVE" ]; then
-    whiptail --title "Operation Successful!" --msgbox "$successtext" 15 80
-  fi
-}
-
-# TODO: Unite with function above
-openhab2_stable_updowngrade() {
-  introtext="You are about to switch over to the stable openHAB 2.1.0 build. When prompted if files should be replaced by newer ones, select Yes. Please be sure to take a full openHAB configuration backup first!"
-  successtext="The stable release of openHAB 2.1.0 is now installed on your system. Please test the correct behavior of your setup. Check the \"openHAB 2.1 Release Notes\" and the official announcements to learn about additons, fixes and changes:\n
-  ➡ http://www.kaikreuzer.de/2017/06/28/openhab21
-  ➡ https://github.com/openhab/openhab-distro/releases/tag/2.1.0"
-  echo -n "$(timestamp) [openHABian] Installing or switching to openHAB 2.1.0 (stable)... "
-
-  if [ -n "$INTERACTIVE" ]; then
-    if ! (whiptail --title "Description, Continue?" --yes-button "Continue" --no-button "Back" --yesno "$introtext" 15 80) then echo "CANCELED"; return 0; fi
-  fi
-
-  echo "deb https://dl.bintray.com/openhab/apt-repo2 stable main" > /etc/apt/sources.list.d/openhab2.list
-  cond_redirect wget -O openhab-key.asc 'https://bintray.com/user/downloadSubjectPublicKey?username=openhab'
-  cond_redirect apt-key add openhab-key.asc
-  if [ $? -ne 0 ]; then echo "FAILED (key)"; exit 1; fi
-  rm -f openhab-key.asc
-  cond_redirect apt update
-  cond_redirect apt -y install openhab2=2.1.0-1
-  if [ $? -ne 0 ]; then echo "FAILED (apt)"; exit 1; fi
-  cond_redirect adduser openhab dialout
-  cond_redirect adduser openhab tty
-  cond_redirect adduser openhab gpio
-  cond_redirect adduser openhab audio
-  cond_redirect dashboard_add_tile openhabiandocs
-  if [ $? -ne 0 ]; then echo "FAILED (dashboard tile)"; fi
-  cond_redirect systemctl daemon-reload
-  cond_redirect systemctl enable openhab2.service
-  if [ $? -eq 0 ]; then echo "OK"; else echo "FAILED"; exit 1; fi
-  cond_redirect systemctl restart openhab2.service || true
-
-  if [ -n "$INTERACTIVE" ]; then
-    whiptail --title "Operation Successful!" --msgbox "$successtext" 15 80
+    whiptail --title "Operation Successful!" --msgbox "$successtext_stable" 15 80
   fi
 }
 
 openhab_shell_interfaces() {
   introtext="The openHAB remote console is a powerful tool for every openHAB user. It allows you too have a deeper insight into the internals of your setup. Further details: http://docs.openhab.org/administration/console.html
-\nThis routine will bind the console to all interfaces and thereby make it available to other devices in your network. Please provide a secure password for this connection (letters and numbers only! default: habopen):"
+\\nThis routine will bind the console to all interfaces and thereby make it available to other devices in your network. Please provide a secure password for this connection (letters and numbers only! default: habopen):"
   failtext="Sadly there was a problem setting up the selected option. Please report this problem in the openHAB community forum or as a openHABian GitHub issue."
   successtext="The openHAB remote console was successfully opened on all interfaces. openHAB has been restarted. You should be able to reach the console via:
-\n'ssh://openhab:<password>@<openhabian-IP> -p 8101'\n
+\\n'ssh://openhab:<password>@<openhabian-IP> -p 8101'\\n
 Please be aware, that the first connection attempt may take a few minutes or may result in a timeout due to key generation."
 
   echo -n "$(timestamp) [openHABian] Binding the openHAB remote console on all interfaces... "
@@ -137,7 +104,7 @@ nano_openhab_syntax() {
   # add nano syntax highlighting
   echo -n "$(timestamp) [openHABian] Adding openHAB syntax to nano editor... "
   cond_redirect wget -O /usr/share/nano/openhab.nanorc https://raw.githubusercontent.com/airix1/openhabnano/master/openhab.nanorc
-  echo -e "\n## openHAB files\ninclude \"/usr/share/nano/openhab.nanorc\"" >> /etc/nanorc
+  echo -e "\\n## openHAB files\\ninclude \"/usr/share/nano/openhab.nanorc\"" >> /etc/nanorc
   echo "OK"
 }
